authorization added to all routes

+++ backend/src/controllers/budgetActivityLogController.js
@@ -2,16 +2,32 @@ import ActivityLog from "../models/BudgetActivityLog.js";
 import BudgetItem from "../models/BudgetItem.js";
 
 // CREATE
-// create an activity log
+/**
+ * create an activity log
+ * @param {*} req 
+ * @param {*} res 
+ * @returns 
+ */
 export async function createActivityLog(req, res) {
   try {
     const { budgetItem, name, amount, activityDate, notes } =
       req.body;
 
     // does budgetItem exist
-    const item = await BudgetItem.findById(budgetItem);
+    const item = await BudgetItem.findById(budgetItem)
+      .populate("monthlyBudget");
 
-    if (!item) return res.status(404).json({ message: "Item not found" });
+    if (!item) {
+      return res.status(404).json({
+        message: "Item not found",
+      });
+    }
+
+    if (!item.monthlyBudget.userProfile.equals(req.profile._id)) {
+      return res.status(403).json({
+        message: "Forbidden",
+      });
+    }
 
     const savedActivityLog = await ActivityLog.create({
       budgetItem,
@@ -38,8 +54,18 @@ export async function createActivityLog(req, res) {
 }
 
 // READ
-// get all activity logs
+/**
+ * get all activity logs
+ * @param {*} req 
+ * @param {*} res 
+ * @returns 
+ */
 export async function getAllActivityLogs(req, res) {
+  // authorization
+  if (process.env.NODE_ENV === "production") {
+    return res.status(403).json({ message: "Forbidden" });
+  }
+
   try {
     const activityLogs = await ActivityLog.find().sort({ createdAt: -1 });
     res.status(200).json({ activityLogs });
@@ -49,12 +75,33 @@ export async function getAllActivityLogs(req, res) {
   }
 }
 
-// get activity log by id
+/**
+ * get activity log by id
+ * @param {*} req 
+ * @param {*} res 
+ * @returns 
+ */
 export async function getActivityLogById(req, res) {
   try {
-    const activityLog = await ActivityLog.findById(req.params.id);
-    if (!activityLog)
-      return res.status(404).json({ message: "Activity log not found" });
+    const activityLog = await ActivityLog.findById(req.params.id)
+      .populate({
+        path: "budgetItem",
+        populate: {
+          path: "monthlyBudget",
+        },
+      });
+
+    if (!activityLog) {
+      return res.status(404).json({
+        message: "Activity log not found",
+      });
+    }
+
+    if (!activityLog.budgetItem.monthlyBudget.userProfile.equals(req.profile._id)) {
+      return res.status(403).json({
+        message: "Forbidden",
+      });
+    }
     res.status(200).json({
       activityLog,
     });
@@ -64,14 +111,35 @@ export async function getActivityLogById(req, res) {
   }
 }
 
-// get activity log by budget item
+/**
+ * get activity log by budget item
+ * @param {*} req 
+ * @param {*} res 
+ * @returns 
+ */
 export async function getActivityLogsByBudget(req, res) {
   try {
-    const activityLogs = await ActivityLog.find({
-      budgetItem: req.params.budgetItemId,
-    }).sort({
-      activityDate: -1,
+    const item = await BudgetItem.findById(req.params.budgetItemId)
+  .populate("monthlyBudget");
+
+  if (!item) {
+    return res.status(404).json({
+      message: "Item not found",
     });
+  }
+
+  if (!item.monthlyBudget.userProfile.equals(req.profile._id)) {
+    return res.status(403).json({
+      message: "Forbidden",
+    });
+  }
+
+  const activityLogs = await ActivityLog.find({
+    budgetItem: item._id,
+  }).sort({
+    activityDate: -1,
+  });
+  
     res.status(200).json({ activityLogs });
   } catch (err) {
     console.error("getActivityLogsByBudget(): ", err);
@@ -80,9 +148,33 @@ export async function getActivityLogsByBudget(req, res) {
 }
 
 // UPDATE
-// update activity log
+/**
+ * update activity log
+ * @param {*} req 
+ * @param {*} res 
+ * @returns 
+ */
 export async function updateActivityLog(req, res) {
   try {
+    const activityLog = await ActivityLog.findById(req.params.id)
+      .populate({
+        path: "budgetItem",
+        populate: {
+          path: "monthlyBudget",
+        },
+      });
+
+    if (!activityLog) {
+      return res.status(404).json({
+        message: "Activity log not found",
+      });
+    }
+
+    if (!activityLog.budgetItem.monthlyBudget.userProfile.equals(req.profile._id)) {
+      return res.status(403).json({
+        message: "Forbidden",
+      });
+    }
     const { name, amount, activityDate, notes } = req.body;
     const updatedActivityLog = await ActivityLog.findByIdAndUpdate(
       req.params.id,
@@ -107,14 +199,42 @@ export async function updateActivityLog(req, res) {
 }
 
 // DELETE
-// delete activity log
+/**
+ * delete activity log
+ * @param {*} req 
+ * @param {*} res 
+ * @returns 
+ */
 export async function deleteActivityLog(req, res) {
   try {
+    const activityLog = await ActivityLog.findById(req.params.id)
+      .populate({
+        path: "budgetItem",
+        populate: {
+          path: "monthlyBudget",
+        },
+      });
+
+    if (!activityLog) {
+      return res.status(404).json({
+        message: "Activity log not found",
+      });
+    }
+
+    if (!activityLog.budgetItem.monthlyBudget.userProfile.equals(req.profile._id)) {
+      return res.status(403).json({
+        message: "Forbidden",
+      });
+    }
+
+
     const deletedActivityLog = await ActivityLog.findByIdAndDelete(
       req.params.id,
     );
+
     if (!deletedActivityLog)
       return res.status(404).json({ message: "Activity log not found" });
+
     res.status(200).json({
       deletedActivityLog,
     });
@@ -127,6 +247,20 @@ export async function deleteActivityLog(req, res) {
 // delete activity log by budget item
 export async function deleteActivityLogByBudget(req, res) {
   try {
+    const item = await BudgetItem.findById(req.params.budgetItemId)
+    .populate("monthlyBudget");
+
+    if (!item) {
+      return res.status(404).json({
+        message: "Item not found",
+      });
+    }
+
+    if (!item.monthlyBudget.userProfile.equals(req.profile._id)) {
+      return res.status(403).json({
+        message: "Forbidden",
+      });
+    }
     const deletedActivityLogs = await ActivityLog.deleteMany({
       budgetItem: req.params.budgetItemId,
     });
